requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
openai
uvicorn
numpy
pillow
tensorflow
torch
transformers
requests